Slim down player class
