+++ src/player.cpp
@@ -90,25 +90,14 @@ static const efftype_id effect_sleep( "sleep" );
 static const itype_id itype_large_repairkit( "large_repairkit" );
 static const itype_id itype_small_repairkit( "small_repairkit" );
 
-static const trait_id trait_CHLOROMORPH( "CHLOROMORPH" );
 static const trait_id trait_DEBUG_BIONIC_POWER( "DEBUG_BIONIC_POWER" );
 static const trait_id trait_DEBUG_HS( "DEBUG_HS" );
-static const trait_id trait_M_SKIN3( "M_SKIN3" );
 static const trait_id trait_NOMAD( "NOMAD" );
 static const trait_id trait_NOMAD2( "NOMAD2" );
 static const trait_id trait_NOMAD3( "NOMAD3" );
-static const trait_id trait_SHELL2( "SHELL2" );
-static const trait_id trait_THRESH_SPIDER( "THRESH_SPIDER" );
-static const trait_id trait_WATERSLEEP( "WATERSLEEP" );
-static const trait_id trait_WEB_SPINNER( "WEB_SPINNER" );
-static const trait_id trait_WEB_WALKER( "WEB_WALKER" );
-static const trait_id trait_WEB_WEAVER( "WEB_WEAVER" );
 
 static const skill_id skill_firstaid( "firstaid" );
 static const skill_id skill_swimming( "swimming" );
-
-static const bionic_id bio_soporific( "bio_soporific" );
-
 
 player::player()
 {
@@ -635,79 +624,6 @@ void player::mend_item( item_location &&obj, bool interactive )
     }
 }
 
-cata::optional<std::list<item>::iterator>
-player::wear( int pos, bool interactive )
-{
-    return wear( item_location( *this, &i_at( pos ) ), interactive );
-}
-
-cata::optional<std::list<item>::iterator>
-player::wear( item_location item_wear, bool interactive )
-{
-    item to_wear = *item_wear;
-    if( is_worn( to_wear ) ) {
-        if( interactive ) {
-            add_msg_player_or_npc( m_info,
-                                   _( "You are already wearing that." ),
-                                   _( "<npcname> is already wearing that." )
-                                 );
-        }
-        return cata::nullopt;
-    }
-    if( to_wear.is_null() ) {
-        if( interactive ) {
-            add_msg_player_or_npc( m_info,
-                                   _( "You don't have that item." ),
-                                   _( "<npcname> doesn't have that item." ) );
-        }
-        return cata::nullopt;
-    }
-
-    bool was_weapon;
-    item to_wear_copy( to_wear );
-    if( &to_wear == &weapon ) {
-        weapon = item();
-        was_weapon = true;
-    } else if( has_item( to_wear ) ) {
-        remove_item( to_wear );
-        was_weapon = false;
-    } else {
-        // item is on the map if this point is reached.
-        item_wear.remove_item();
-        was_weapon = false;
-    }
-
-    const bool item_one_per_layer = to_wear_copy.has_flag( flag_id( "ONE_PER_LAYER" ) );
-    for( const item &worn_item : worn ) {
-        const cata::optional<side> sidedness_conflict = to_wear_copy.covers_overlaps( worn_item );
-        if( sidedness_conflict && ( item_one_per_layer ||
-                                    worn_item.has_flag( flag_id( "ONE_PER_LAYER" ) ) ) ) {
-            // we can assume both isn't an option because it'll be caught in can_wear
-            if( *sidedness_conflict == side::LEFT ) {
-                to_wear_copy.set_side( side::RIGHT );
-            } else {
-                to_wear_copy.set_side( side::LEFT );
-            }
-        }
-    }
-
-    auto result = wear_item( to_wear_copy, interactive );
-    if( !result ) {
-        if( was_weapon ) {
-            weapon = to_wear_copy;
-        } else {
-            i_add( to_wear_copy );
-        }
-        return cata::nullopt;
-    }
-
-    if( was_weapon ) {
-        get_event_bus().send<event_type::character_wields_item>( getID(), weapon.typeId() );
-    }
-
-    return result;
-}
-
 template <typename T>
 bool player::can_lift( const T &obj ) const
 {
@@ -722,92 +638,6 @@ bool player::can_lift( const T &obj ) const
 }
 template bool player::can_lift<item>( const item &obj ) const;
 template bool player::can_lift<vehicle>( const vehicle &obj ) const;
-
-void player::use_wielded()
-{
-    use( -1 );
-}
-
-void player::use( int inventory_position )
-{
-    item &used = i_at( inventory_position );
-    item_location loc = item_location( *this, &used );
-
-    use( loc );
-}
-
-void player::use( item_location loc, int pre_obtain_moves )
-{
-    // if -1 is passed in we don't want to change moves at all
-    if( pre_obtain_moves == -1 ) {
-        pre_obtain_moves = moves;
-    }
-    if( !loc ) {
-        add_msg( m_info, _( "You do not have that item." ) );
-        moves = pre_obtain_moves;
-        return;
-    }
-
-    item &used = *loc;
-    last_item = used.typeId();
-
-    if( used.is_tool() ) {
-        if( !used.type->has_use() ) {
-            add_msg_if_player( _( "You can't do anything interesting with your %s." ), used.tname() );
-            moves = pre_obtain_moves;
-            return;
-        }
-        invoke_item( &used, loc.position(), pre_obtain_moves );
-
-    } else if( used.type->can_use( "DOGFOOD" ) ||
-               used.type->can_use( "CATFOOD" ) ||
-               used.type->can_use( "BIRDFOOD" ) ||
-               used.type->can_use( "CATTLEFODDER" ) ) { // NOLINT(bugprone-branch-clone)
-        invoke_item( &used, loc.position(), pre_obtain_moves );
-
-    } else if( !used.is_craft() && ( used.is_medication() || ( !used.type->has_use() &&
-                                     used.is_food() ) ) ) {
-
-        if( used.is_medication() && !can_use_heal_item( used ) ) {
-            add_msg_if_player( m_bad, _( "Your biology is not compatible with that healing item." ) );
-            moves = pre_obtain_moves;
-            return;
-        }
-
-        if( avatar *u = as_avatar() ) {
-            const ret_val<edible_rating> ret = u->will_eat( used, true );
-            if( !ret.success() ) {
-                moves = pre_obtain_moves;
-                return;
-            }
-            u->assign_activity( player_activity( consume_activity_actor( item_location( *u, &used ) ) ) );
-        } else  {
-            const time_duration &consume_time = get_consume_time( used );
-            moves -= to_moves<int>( consume_time );
-            consume( used );
-        }
-    } else if( used.is_book() ) {
-        // TODO: Handle this with dynamic dispatch.
-        if( avatar *u = as_avatar() ) {
-            u->read( loc );
-        }
-    } else if( used.type->has_use() ) {
-        invoke_item( &used, loc.position(), pre_obtain_moves );
-    } else if( used.has_flag( flag_SPLINT ) ) {
-        ret_val<bool> need_splint = can_wear( *loc );
-        if( need_splint.success() ) {
-            wear_item( used );
-            loc.remove_item();
-        } else {
-            add_msg( m_info, need_splint.str() );
-        }
-    } else if( used.is_relic() ) {
-        invoke_item( &used, loc.position(), pre_obtain_moves );
-    } else {
-        add_msg( m_info, _( "You can't do anything interesting with your %s." ), used.tname() );
-        moves = pre_obtain_moves;
-    }
-}
 
 bool player::gunmod_remove( item &gun, item &mod )
 {
@@ -943,207 +773,6 @@ void player::toolmod_add( item_location tool, item_location mod )
     assign_activity( activity_id( "ACT_TOOLMOD_ADD" ), 1, -1 );
     activity.targets.emplace_back( std::move( tool ) );
     activity.targets.emplace_back( std::move( mod ) );
-}
-
-void player::try_to_sleep( const time_duration &dur )
-{
-    map &here = get_map();
-    const optional_vpart_position vp = here.veh_at( pos() );
-    const trap &trap_at_pos = here.tr_at( pos() );
-    const ter_id ter_at_pos = here.ter( pos() );
-    const furn_id furn_at_pos = here.furn( pos() );
-    bool plantsleep = false;
-    bool fungaloid_cosplay = false;
-    bool websleep = false;
-    bool webforce = false;
-    bool websleeping = false;
-    bool in_shell = false;
-    bool watersleep = false;
-    if( has_trait( trait_CHLOROMORPH ) ) {
-        plantsleep = true;
-        if( ( ter_at_pos == t_dirt || ter_at_pos == t_pit ||
-              ter_at_pos == t_dirtmound || ter_at_pos == t_pit_shallow ||
-              ter_at_pos == t_grass ) && !vp &&
-            furn_at_pos == f_null ) {
-            add_msg_if_player( m_good, _( "You relax as your roots embrace the soil." ) );
-        } else if( vp ) {
-            add_msg_if_player( m_bad, _( "It's impossible to sleep in this wheeled pot!" ) );
-        } else if( furn_at_pos != f_null ) {
-            add_msg_if_player( m_bad,
-                               _( "The humans' furniture blocks your roots.  You can't get comfortable." ) );
-        } else { // Floor problems
-            add_msg_if_player( m_bad, _( "Your roots scrabble ineffectively at the unyielding surface." ) );
-        }
-    } else if( has_trait( trait_M_SKIN3 ) ) {
-        fungaloid_cosplay = true;
-        if( here.has_flag_ter_or_furn( "FUNGUS", pos() ) ) {
-            add_msg_if_player( m_good,
-                               _( "Our fibers meld with the ground beneath us.  The gills on our neck begin to seed the air with spores as our awareness fades." ) );
-        }
-    }
-    if( has_trait( trait_WEB_WALKER ) ) {
-        websleep = true;
-    }
-    // Not sure how one would get Arachnid w/o web-making, but Just In Case
-    if( has_trait( trait_THRESH_SPIDER ) && ( has_trait( trait_WEB_SPINNER ) ||
-            ( has_trait( trait_WEB_WEAVER ) ) ) ) {
-        webforce = true;
-    }
-    if( websleep || webforce ) {
-        int web = here.get_field_intensity( pos(), fd_web );
-        if( !webforce ) {
-            // At this point, it's kinda weird, but surprisingly comfy...
-            if( web >= 3 ) {
-                add_msg_if_player( m_good,
-                                   _( "These thick webs support your weight, and are strangely comfortable…" ) );
-                websleeping = true;
-            } else if( web > 0 ) {
-                add_msg_if_player( m_info,
-                                   _( "You try to sleep, but the webs get in the way.  You brush them aside." ) );
-                here.remove_field( pos(), fd_web );
-            }
-        } else {
-            // Here, you're just not comfortable outside a nice thick web.
-            if( web >= 3 ) {
-                add_msg_if_player( m_good, _( "You relax into your web." ) );
-                websleeping = true;
-            } else {
-                add_msg_if_player( m_bad,
-                                   _( "You try to sleep, but you feel exposed and your spinnerets keep twitching." ) );
-                add_msg_if_player( m_info, _( "Maybe a nice thick web would help you sleep." ) );
-            }
-        }
-    }
-    if( has_active_mutation( trait_SHELL2 ) ) {
-        // Your shell's interior is a comfortable place to sleep.
-        in_shell = true;
-    }
-    if( has_trait( trait_WATERSLEEP ) ) {
-        if( underwater ) {
-            add_msg_if_player( m_good,
-                               _( "You lay beneath the waves' embrace, gazing up through the water's surface…" ) );
-            watersleep = true;
-        } else if( here.has_flag_ter( "SWIMMABLE", pos() ) ) {
-            add_msg_if_player( m_good, _( "You settle into the water and begin to drowse…" ) );
-            watersleep = true;
-        }
-    }
-    if( !plantsleep && ( furn_at_pos.obj().comfort > static_cast<int>( comfort_level::neutral ) ||
-                         ter_at_pos == t_improvised_shelter ||
-                         trap_at_pos.comfort > static_cast<int>( comfort_level::neutral ) ||
-                         in_shell || websleeping || watersleep ||
-                         vp.part_with_feature( "SEAT", true ) ||
-                         vp.part_with_feature( "BED", true ) ) ) {
-        add_msg_if_player( m_good, _( "This is a comfortable place to sleep." ) );
-    } else if( !plantsleep && !fungaloid_cosplay && !watersleep ) {
-        if( !vp && ter_at_pos != t_floor ) {
-            add_msg_if_player( ter_at_pos.obj().movecost <= 2 ?
-                               _( "It's a little hard to get to sleep on this %s." ) :
-                               _( "It's hard to get to sleep on this %s." ),
-                               ter_at_pos.obj().name() );
-        } else if( vp ) {
-            if( vp->part_with_feature( VPFLAG_AISLE, true ) ) {
-                add_msg_if_player(
-                    //~ %1$s: vehicle name, %2$s: vehicle part name
-                    _( "It's a little hard to get to sleep on this %2$s in %1$s." ),
-                    vp->vehicle().disp_name(),
-                    vp->part_with_feature( VPFLAG_AISLE, true )->part().name( false ) );
-            } else {
-                add_msg_if_player(
-                    //~ %1$s: vehicle name
-                    _( "It's hard to get to sleep in %1$s." ),
-                    vp->vehicle().disp_name() );
-            }
-        }
-    }
-    add_msg_if_player( _( "You start trying to fall asleep." ) );
-    if( has_active_bionic( bio_soporific ) ) {
-        bio_soporific_powered_at_last_sleep_check = has_power();
-        if( bio_soporific_powered_at_last_sleep_check ) {
-            // The actual bonus is applied in sleep_spot( p ).
-            add_msg_if_player( m_good, _( "Your soporific inducer starts working its magic." ) );
-        } else {
-            add_msg_if_player( m_bad, _( "Your soporific inducer doesn't have enough power to operate." ) );
-        }
-    }
-    assign_activity( player_activity( try_sleep_activity_actor( dur ) ) );
-}
-
-bool player::wield_contents( item &container, item *internal_item, bool penalties, int base_cost )
-{
-    // if index not specified and container has multiple items then ask the player to choose one
-    if( internal_item == nullptr ) {
-        std::vector<std::string> opts;
-        std::list<item *> container_contents = container.all_items_top();
-        std::transform( container_contents.begin(), container_contents.end(),
-        std::back_inserter( opts ), []( const item * elem ) {
-            return elem->display_name();
-        } );
-        if( opts.size() > 1 ) {
-            int pos = uilist( _( "Wield what?" ), opts );
-            if( pos < 0 ) {
-                return false;
-            }
-            internal_item = *std::next( container_contents.begin(), pos );
-        } else {
-            internal_item = container_contents.front();
-        }
-    }
-
-    if( !container.has_item( *internal_item ) ) {
-        debugmsg( "Tried to wield non-existent item from container (player::wield_contents)" );
-        return false;
-    }
-
-    const ret_val<bool> ret = can_wield( *internal_item );
-    if( !ret.success() ) {
-        add_msg_if_player( m_info, "%s", ret.c_str() );
-        return false;
-    }
-
-    int mv = 0;
-
-    if( has_wield_conflicts( *internal_item ) ) {
-        if( !unwield() ) {
-            return false;
-        }
-        inv->unsort();
-    }
-
-    // for holsters, we should not include the cost of wielding the holster itself
-    // The cost of wielding the holster was already added earlier in avatar_action::use_item.
-    // As we couldn't make sure back then what action was going to be used, we remove the cost now.
-    item_location il = item_location( *this, &container );
-    mv -= il.obtain_cost( *this );
-    mv += item_retrieve_cost( *internal_item, container, penalties, base_cost );
-
-    if( internal_item->stacks_with( weapon, true ) ) {
-        weapon.combine( *internal_item );
-    } else {
-        weapon = std::move( *internal_item );
-    }
-    container.remove_item( *internal_item );
-    container.on_contents_changed();
-
-    inv->update_invlet( weapon );
-    inv->update_cache_with_item( weapon );
-    last_item = weapon.typeId();
-
-    moves -= mv;
-
-    weapon.on_wield( *this );
-
-    get_event_bus().send<event_type::character_wields_item>( getID(), weapon.typeId() );
-
-    return true;
-}
-
-void player::store( item &container, item &put, bool penalties, int base_cost,
-                    item_pocket::pocket_type pk_type )
-{
-    moves -= item_store_cost( put, container, penalties, base_cost );
-    container.put_in( i_rem( &put ), pk_type );
-    calc_encumbrance();
 }
 
 nc_color encumb_color( int level )
