+++ src/player.h
@@ -115,16 +115,6 @@ class player : public Character
          */
         int climbing_cost( const tripoint &from, const tripoint &to ) const;
 
-        /** Used for eating object at a location. Removes item if all of it was consumed.
-        *   @returns trinary enum NONE, SOME or ALL amount consumed.
-        */
-        trinary consume( item_location loc, bool force = false, bool refuel = false );
-
-        /** Used for eating a particular item that doesn't need to be in inventory.
-         *  @returns trinary enum NONE, SOME or ALL (doesn't remove).
-         */
-        trinary consume( item &target, bool force = false, bool refuel = false );
-
         /** Check player strong enough to lift an object unaided by equipment (jacks, levers etc) */
         template <typename T> bool can_lift( const T &obj ) const;
 
@@ -135,44 +125,8 @@ class player : public Character
          */
         void mend_item( item_location &&obj, bool interactive = true );
 
-        /** Wear item; returns false on fail. If interactive is false, don't alert the player or drain moves on completion. */
-        cata::optional<std::list<item>::iterator>
-        wear( int pos, bool interactive = true );
-
-        /** Wear item; returns false on fail. If interactive is false, don't alert the player or drain moves on completion.
-        * @param item_wear item_location of item to be worn.
-        * @param interactive Alert player and drain moves if true.
-        */
-        cata::optional<std::list<item>::iterator>
-        wear( item_location item_wear, bool interactive = true );
-
-        /**
-         * Try to wield a contained item consuming moves proportional to weapon skill and volume.
-         * @param container Container containing the item to be wielded
-         * @param internal_item reference to contained item to wield.
-         * @param penalties Whether item volume and temporary effects (e.g. GRABBED, DOWNED) should be considered.
-         * @param base_cost Cost due to storage type.
-         */
-        bool wield_contents( item &container, item *internal_item = nullptr, bool penalties = true,
-                             int base_cost = INVENTORY_HANDLING_PENALTY );
-        /**
-         * Stores an item inside another consuming moves proportional to weapon skill and volume
-         * @param container Container in which to store the item
-         * @param put Item to add to the container
-         * @param penalties Whether item volume and temporary effects (e.g. GRABBED, DOWNED) should be considered.
-         * @param base_cost Cost due to storage type.
-         */
-        void store( item &container, item &put, bool penalties = true,
-                    int base_cost = INVENTORY_HANDLING_PENALTY,
-                    item_pocket::pocket_type pk_type = item_pocket::pocket_type::CONTAINER );
         /** Draws the UI and handles player input for the armor re-ordering window */
         void sort_armor();
-        /** Uses a tool */
-        void use( int inventory_position );
-        /** Uses a tool at location */
-        void use( item_location loc, int pre_obtain_moves = -1 );
-        /** Uses the current wielded weapon */
-        void use_wielded();
 
         /**
          * Starts activity to remove gunmod after unloading any contained ammo.
@@ -185,9 +139,6 @@ class player : public Character
 
         /** Starts activity to install toolmod */
         void toolmod_add( item_location tool, item_location mod );
-
-        /** Handles sleep attempts by the player, starts ACT_TRY_SLEEP activity */
-        void try_to_sleep( const time_duration &dur );
 
         // ---------------VALUES-----------------
         // Relative direction of a grab, add to posx, posy to get the coordinates of the grabbed thing.
